refactor(Goal/feat-goal): interface 구현하도록 컨트롤러와 서비스 수정 #1
